+++ crate/cli/src/tests/kms/elliptic_curve/mod.rs
@@ -1,3 +1,4 @@
 pub(super) mod create_key_pair;
 #[cfg(feature = "non-fips")]
 pub(super) mod encrypt_decrypt;
+mod sign_verify;